test: shard tui gateway s4 clusters
