+++ tests/test_tui_gateway_server.py
@@ -10172,42 +10172,6 @@ def test_wait_agent_for_prompt_expires_at_cap(monkeypatch):
         server._sessions.pop("sid", None)
 
 
-def test_clear_pending_without_sid_clears_all():
-    """_clear_pending(None) is the shutdown path — must still release
-    every pending prompt regardless of owning session."""
-    ev1, ev2, ev3 = threading.Event(), threading.Event(), threading.Event()
-    server._pending["a"] = ("sid_x", ev1)
-    server._pending["b"] = ("sid_y", ev2)
-    server._pending["c"] = ("sid_z", ev3)
-    try:
-        server._clear_pending(None)
-        assert ev1.is_set() and ev2.is_set() and ev3.is_set()
-    finally:
-        for key in ("a", "b", "c"):
-            server._pending.pop(key, None)
-            server._answers.pop(key, None)
-
-
-def test_respond_unpacks_sid_tuple_correctly():
-    """After the (sid, Event) tuple change, _respond must still work."""
-    ev = threading.Event()
-    server._pending["rid-x"] = ("sid_x", ev)
-    try:
-        resp = server.handle_request(
-            {
-                "id": "1",
-                "method": "clarify.respond",
-                "params": {"request_id": "rid-x", "answer": "the answer"},
-            }
-        )
-        assert resp.get("result")
-        assert ev.is_set()
-        assert server._answers.get("rid-x") == "the answer"
-    finally:
-        server._pending.pop("rid-x", None)
-        server._answers.pop("rid-x", None)
-
-
 # ---------------------------------------------------------------------------
 # /model switch and other agent-mutating commands must reject while the
 # session is running.  agent.switch_model() mutates self.model, self.provider,
@@ -10217,97 +10181,6 @@ def test_respond_unpacks_sid_tuple_correctly():
 # (session["pending_model_switch"]) and _apply_pending_model_switch applies it
 # on the turn thread at the next turn start, where nothing is in flight.
 # ---------------------------------------------------------------------------
-
-
-def test_config_set_model_defers_while_running(monkeypatch):
-    """/model via config.set queues the pick during an in-flight turn instead
-    of rejecting or racing the worker thread."""
-    seen = {"called": False}
-
-    def _fake_apply(sid, session, raw, **_kwargs):
-        seen["called"] = True
-        return {"value": raw, "warning": ""}
-
-    monkeypatch.setattr(server, "_apply_model_switch", _fake_apply)
-
-    server._sessions["sid"] = _session(running=True)
-    try:
-        resp = server.handle_request(
-            {
-                "id": "1",
-                "method": "config.set",
-                "params": {
-                    "session_id": "sid",
-                    "key": "model",
-                    "value": "anthropic/claude-sonnet-4.6",
-                },
-            }
-        )
-        assert not resp.get("error")
-        result = resp["result"]
-        assert result["deferred"] is True
-        assert result["value"] == "anthropic/claude-sonnet-4.6"
-        assert not seen["called"], (
-            "_apply_model_switch ran mid-turn — would race the worker thread "
-            "reading agent.model / agent.client; it must defer to turn start"
-        )
-        pending = server._sessions["sid"].get("pending_model_switch")
-        assert pending and pending["raw"] == "anthropic/claude-sonnet-4.6"
-    finally:
-        server._sessions.pop("sid", None)
-
-
-def test_apply_pending_model_switch_runs_queued_pick(monkeypatch):
-    """The queued pick is consumed once, on the turn thread, via
-    _apply_model_switch — and cleared so it can't re-fire next turn."""
-    calls = []
-
-    def _fake_apply(sid, session, raw, **kwargs):
-        calls.append(raw)
-        return {"value": raw, "warning": "", "confirm_required": False}
-
-    monkeypatch.setattr(server, "_apply_model_switch", _fake_apply)
-
-    session = _session(running=False)
-    session["agent"] = object()
-    session["pending_model_switch"] = {
-        "raw": "anthropic/claude-sonnet-4.6",
-        "confirm_expensive_model": False,
-    }
-
-    server._apply_pending_model_switch("sid", session)
-    assert calls == ["anthropic/claude-sonnet-4.6"]
-    assert "pending_model_switch" not in session
-
-    # Idempotent: a second turn start with nothing queued is a no-op.
-    server._apply_pending_model_switch("sid", session)
-    assert calls == ["anthropic/claude-sonnet-4.6"]
-
-
-def test_config_set_model_allowed_when_idle(monkeypatch):
-    """Regression guard: idle sessions can still switch models."""
-    seen = {"called": False}
-
-    def _fake_apply(sid, session, raw, **_kwargs):
-        seen["called"] = True
-        return {"value": "newmodel", "warning": ""}
-
-    monkeypatch.setattr(server, "_apply_model_switch", _fake_apply)
-
-    server._sessions["sid"] = _session(running=False)
-    try:
-        resp = server.handle_request(
-            {
-                "id": "1",
-                "method": "config.set",
-                "params": {"session_id": "sid", "key": "model", "value": "newmodel"},
-            }
-        )
-        assert resp.get("result")
-        assert resp["result"]["value"] == "newmodel"
-        assert seen["called"]
-    finally:
-        server._sessions.pop("sid", None)
 
 
 def test_mirror_slash_side_effects_rejects_mutating_commands_while_running(monkeypatch):
